Extend jpaSchemaGen task with filtering of Entity classes by package
Add a new config property `includeOnlyPackages` to the jpaSchemaGen task.
If set to a non-empty list, it will filter the entity classes listed in
the persistence.xml using this package white-list.
An entity class is included if one of the `includeOnlyPackages` is a
prefix of its fully qualified name.

This feature has two limitations:
 - it only works if all classes are explicitly listed in the persistence.xml
 - the implementation is a bit on the hack-y side; for maximum stability
   the more direct implementation is preferred if the `includeOnlyPackages` feature is not configured.